pyproject.toml
[build-system]
requires = ["setuptools>=74.1.1"]
build-backend = "setuptools.build_meta"

[project]
name = "soundingcenter"
version = "1.3.4"
description = "A python lib to use the sounding center api"
readme = "README.md"
requires-python = ">=3.14"
license = { file = "LICENSE" }
authors = [{ name = "Niklas Wildenburg", email = "niklas.wildenburg@graw.de" }]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = ["requests~=2.32.5"]

[dependency-groups]
"dev" = ["Faker~=40.8.0", "python-dotenv~=1.2.2"]

[project.urls]
"Homepage" = "https://github.com/GrawRadiosondes/sounding-center-python-api"
"Bug Tracker" = "https://github.com/GrawRadiosondes/sounding-center-python-api/issues"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::urllib3.exceptions.InsecureRequestWarning"
]
